+++ src/Correntra.Desktop/Views/SettingsWindow.axaml.cs
@@ -1,8 +1,8 @@
+using System.Diagnostics;
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Styling;
-using System.Diagnostics;
 using Correntra.Desktop.Services;
 using Correntra.Desktop.ViewModels;